refactor diffusion triton kernels
