+++ python/sglang/multimodal_gen/runtime/layers/elementwise.py
@@ -1,7 +1,7 @@
 import torch
 
+from sglang.jit_kernel.diffusion.triton.scale_shift import fuse_scale_shift_kernel
 from sglang.multimodal_gen.runtime.layers.custom_op import CustomOp
-from sglang.multimodal_gen.runtime.layers.triton_ops import fuse_scale_shift_kernel
 
 
 class MulAdd(CustomOp):
