+++ python/sglang/multimodal_gen/runtime/layers/rotary_embedding.py
@@ -32,9 +32,9 @@ from typing import Any, Optional, Tuple
 
 import torch
 
+from sglang.jit_kernel.diffusion.triton.rotary import apply_rotary_embedding
 from sglang.multimodal_gen.runtime.distributed.parallel_state import get_sp_group
 from sglang.multimodal_gen.runtime.layers.custom_op import CustomOp
-from sglang.multimodal_gen.runtime.layers.triton_ops import apply_rotary_embedding
 from sglang.multimodal_gen.runtime.utils.logging_utils import init_logger
 
 logger = init_logger(__name__)
